# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Library Management System API",
    "description": "A **Mini Library Management System** built with FastAPI, PostgreSQL, and OpenAI.\n\n## Authentication\n\nAll protected endpoints require a **Bearer JWT** obtained through the OAuth login flow:\n\n1. Open `GET /api/v1/auth/login/google` (or `/github`) in your **browser**.\n2. Complete the OAuth consent screen.\n3. You are redirected to the frontend with `?token=<jwt>` in the URL.\n4. Copy the token and paste it into the **Authorize** dialog (🔒 button above).\n\nEvery protected endpoint then works directly from this Swagger UI.\n\n## Roles & Permissions\n\n| Role | Capabilities |\n|------|--------------|\n| **Admin** | Full access including user management |\n| **Librarian** | Create / edit / delete books; manage all loans |\n| **Member** | Browse and search books; borrow and return **own** loans |\n\nNew OAuth users start as **Member**. An Admin must promote them via `PATCH /api/v1/admin/users/{id}/role`.\n\n## AI Features\n\nAll three AI endpoints degrade gracefully — they never return an error due to missing AI configuration. Check the `source` field (`\"openai\"` vs `\"fallback\"`) to see which path was taken.\n",
    "contact": {
      "name": "Library Management System",
      "url": "https://github.com/maherabbass/Library-Management-System"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/loans/checkout": {
      "post": {
        "tags": [
          "loans"
        ],
        "summary": "Borrow a book",
        "description": "Creates a new loan and marks the book as `BORROWED`.\n\n**Business rules:**\n- The book must currently have status `AVAILABLE`. If it is already `BORROWED`, the request fails with `409 Conflict`.\n- There is no limit on how many books a single user can borrow simultaneously.\n- Any authenticated user (Member, Librarian, or Admin) can borrow a book.\n\n**Concurrency:** the endpoint uses `SELECT ... FOR UPDATE` to prevent two users from borrowing the same book simultaneously.\n\n**Requires:** any authenticated user.",
        "operationId": "checkout_endpoint_api_v1_loans_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The new loan record with status `OUT`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "404": {
            "description": "Book not found."
          },
          "409": {
            "description": "Conflict — the book is already borrowed by another user."
          },
          "422": {
            "description": "Validation error — `book_id` must be a valid UUID."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/loans/return": {
      "post": {
        "tags": [
          "loans"
        ],
        "summary": "Return a book",
        "description": "Closes an active loan and marks the book as `AVAILABLE` again.\n\n**Business rules:**\n- The loan must have status `OUT`. Already-returned loans return `404`.\n- **Members** may only return their **own** loans — supplying another user's loan ID returns `403 Forbidden`.\n- **Librarians** and **Admins** may return any loan.\n\n**Requires:** any authenticated user.",
        "operationId": "return_endpoint_api_v1_loans_return_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The closed loan record with status `RETURNED` and `returned_at` timestamp.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "403": {
            "description": "Forbidden — Members may only return their own loans."
          },
          "404": {
            "description": "Loan not found or already returned."
          },
          "422": {
            "description": "Validation error — `loan_id` must be a valid UUID."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/loans": {
      "get": {
        "tags": [
          "loans"
        ],
        "summary": "List loans",
        "description": "Returns a paginated list of loans.\n\n**Role-based filtering:**\n- **Members** see only their **own** loans.\n- **Librarians** and **Admins** see **all** loans across all users.\n\nResults are ordered by `checked_out_at` descending (most recent first).\n\n**Requires:** any authenticated user.",
        "operationId": "list_loans_endpoint_api_v1_loans_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based).",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (1–100).",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page (1–100)."
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated loan list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CheckoutRequest": {
        "properties": {
          "book_id": {
            "type": "string",
            "format": "uuid",
            "title": "Book Id",
            "description": "UUID of the book to borrow. The book must have status `AVAILABLE`.",
            "examples": [
              "3fa85f64-5717-4562-b3fc-2c963f66afa6"
            ]
          }
        },
        "type": "object",
        "required": [
          "book_id"
        ],
        "title": "CheckoutRequest",
        "example": {
          "book_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoanListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LoanResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Loans on the current page."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of loans matching the current query."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "LoanListResponse",
        "example": {
          "items": [],
          "total": 0
        }
      },
      "LoanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique loan identifier (UUID v4)."
          },
          "book_id": {
            "type": "string",
            "format": "uuid",
            "title": "Book Id",
            "description": "UUID of the borrowed book."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id",
            "description": "UUID of the user who borrowed the book."
          },
          "checked_out_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked Out At",
            "description": "UTC timestamp when the book was checked out."
          },
          "returned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned At",
            "description": "UTC timestamp when the book was returned. `null` if still checked out."
          },
          "status": {
            "$ref": "#/components/schemas/LoanStatus",
            "description": "Loan status: `OUT` — book is still borrowed; `RETURNED` — book has been returned."
          }
        },
        "type": "object",
        "required": [
          "id",
          "book_id",
          "user_id",
          "checked_out_at",
          "status"
        ],
        "title": "LoanResponse",
        "example": {
          "book_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "checked_out_at": "2024-01-20T09:00:00Z",
          "id": "8a1bc234-9876-4def-b3fc-1a2b3c4d5e6f",
          "status": "OUT",
          "user_id": "1b2c3d4e-5f6a-7b8c-9d0e-1f2a3b4c5d6e"
        }
      },
      "LoanStatus": {
        "type": "string",
        "enum": [
          "OUT",
          "RETURNED"
        ],
        "title": "LoanStatus"
      },
      "ReturnRequest": {
        "properties": {
          "loan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Loan Id",
            "description": "UUID of the active loan to close. Members may only supply their **own** loan IDs. Librarians and Admins may supply any loan ID.",
            "examples": [
              "8a1bc234-9876-4def-b3fc-1a2b3c4d5e6f"
            ]
          }
        },
        "type": "object",
        "required": [
          "loan_id"
        ],
        "title": "ReturnRequest",
        "example": {
          "loan_id": "8a1bc234-9876-4def-b3fc-1a2b3c4d5e6f"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from `GET /api/v1/auth/callback/{provider}`. Obtain it by completing the OAuth login flow and copying the `token` query parameter from the redirect URL."
      }
    }
  },
...
}